requirements_utf8.txt
absl-py==2.1.0
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio-sdk==0.3.8
alembic==1.16.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
astunparse==1.6.3
asyncpg==0.31.0
atomicwrites==1.4.1
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
av==15.1.0
bark==0.1.5
bce-python-sdk==0.9.55
bcrypt==4.3.0
beautifulsoup4==4.13.5
billiard==4.2.4
black==26.1.0
blinker==1.9.0
boto3==1.37.17
botocore==1.37.17
Brotli==1.1.0
cachetools==5.5.2
celery==5.6.2
certifi==2025.6.15
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.1
crispy-bootstrap5==0.7
cryptography==46.0.3
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
ctranslate2==4.6.0
cycler==0.12.1
datasets==4.1.1
decorator==4.4.2
deepface==0.0.96
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.0
distro==1.9.0
Django==4.2.11
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==23.5
django-jet-reboot==1.3.5
django-model-utils==4.3.1
django-modeltranslation==0.19.9
djangorestframework==3.14.0
dnspython==2.8.0
ecdsa==0.19.1
edge-tts==7.2.3
einops==0.8.1
einops-exts==0.0.4
email-to==0.1.0
email-validator==2.3.0
encodec==0.1.1
et_xmlfile==2.0.0
facenet-pytorch==2.5.3
fastapi==0.116.2
faster-whisper==1.2.0
fer==25.10.3
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
flake8==7.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.56.0
fpdf2==2.8.5
frozenlist==1.7.0
fsspec==2025.3.0
funcy==2.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gopay==2.0.1
gradio==5.46.1
gradio_client==1.13.1
greenlet==3.2.3
groovy==0.1.2
grpcio==1.71.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imagesize==1.4.1
inflate64==1.0.3
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.0
keyboard==0.13.5
kiwisolver==1.4.8
kombu==5.6.2
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.29
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
librt==0.7.8
lightecc==0.0.3
lightphe==0.0.19
llvmlite==0.45.0
loguru==0.7.3
lxml==5.4.0
lz4==4.4.3
Mako==1.3.10
Markdown==3.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
modelscope==1.32.0
monotonic_align @ git+https://github.com/resemble-ai/monotonic_align.git@c6e5e6cb19882164027eb6e35118e841eed9298e
more-itertools==10.7.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.6.4
multiprocess==0.70.16
multivolumefile==0.2.3
munch==4.0.0
mypy==1.19.1
mypy_extensions==1.1.0
namex==0.0.8
narwhals==2.5.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.0
numpy==2.1.1
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.22.1
onnxruntime-gpu==1.22.0
onnxslim==0.1.77
openai==1.108.1
openai-whisper==20250625
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
orjson==3.11.3
oscrypto==1.3.0
packaging==24.2
paddleocr==3.3.2
paddlex==3.3.10
pandas==2.2.3
passlib==1.7.4
pathspec==1.0.3
pillow==10.4.0
pillow_heif==0.22.0
platformdirs==4.4.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
pooch==1.8.2
premailer==3.10.0
prettytable==3.17.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.11
py==1.11.0
py-cpuinfo==9.0.0
py7zr==1.0.0
pyarrow==21.0.0
pyasn1==0.6.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybboxes==0.1.6
pybcj==1.0.6
pyclipper==1.4.0
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.4.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==6.4.2
pypdfium2==5.1.0
pyperclip==1.9.0
pyppmd==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==0.26.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.4.0
pytweening==1.2.0
pytz==2022.7
PyYAML==6.0.2
pyzstd==0.17.0
redis==5.3.1
referencing==0.36.2
regex==2025.9.18
reportlab==4.0.4
requests==2.32.4
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==13.9.4
roboflow==1.1.58
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
ruff==0.13.1
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.6.2
sahi==0.11.22
scikit-learn==1.7.2
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==70.2.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlparse==0.5.5
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.48.0
streamlit==1.49.1
stripe==5.5.0
svglib==1.5.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==2.5.0
terminaltables==3.1.10
texttable==1.7.0
tf_keras==2.20.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.5.1
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.7.1+cu118
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
tts_arabic @ git+https://github.com/nipponjo/tts_arabic.git@b3293217b7e5d889f6db8c28f98e8ecc476274fa
typer==0.19.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
ultralytics==8.3.94
ultralytics-thop==2.0.14
Unidecode==1.4.0
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.36.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.2.0
win32_setctime==1.2.0
wrapt==1.17.2
xhtml2pdf==0.2.15
xxhash==3.5.0
yarl==1.20.1
yolov5==7.0.14
zstandard==0.25.0
